feat: add broker tests for feed item transforms
