+++ packages/openactive-broker-microservice/package.json
@@ -7,7 +7,7 @@
     "start": "node app.js",
     "auto-restart": "./auto-restart.sh",
     "validate-feeds": "node app.js --validate-only",
-    "test": "npm run lint && tsc",
+    "test": "npm run lint && tsc && jest",
     "debug": "node --inspect app.js",
     "lint": "eslint \"*.js\" \"src/**/*.js\"",
     "lint-fix": "npm run lint -- --fix"
@@ -48,15 +48,23 @@
     "@types/cli-progress": "3.9.0",
     "@types/express": "^4.17.9",
     "@types/fs-extra": "^9.0.13",
+    "@types/jest": "^29.5.4",
     "@types/lodash": "^4.14.175",
     "@types/node": "^14.6.0",
     "@types/ramda": "0.27.37",
     "eslint": "^7.24.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-import": "^2.22.1",
-    "typescript": "^3.9.7"
+    "jest": "^29.7.0",
+    "typescript": "^5.2.2"
   },
   "engines": {
     "node": "14.16.0"
+  },
+  "jest": {
+    "testEnvironment": "node",
+    "testMatch": [
+      "**/test/**/*-test.[tj]s?(x)"
+    ]
   }
 }
